Fix unstable specialization syntax future-compatibility warning.
